Separate main() into its own function in lib.rs
