+++ .gitignore
@@ -1,2 +1,5 @@
 /target
 /Cargo.lock
+.DS_Store
+.vscode/launch.json
+*.metallib
